printf on 4S. RAM loading restored. _load_ptr bug fixed. UART0 IER bug fixed.

+++ runtime/include/flipper/error.h
@@ -18,7 +18,7 @@
 	}
 
 /* Describes a type used to contain libflipper error codes. */
-typedef uint32_t lf_error_t;
+typedef uint8_t lf_error_t;
 
 /* Enumerate all of the error codes. */
 enum {
